userId for db

+++ app/_lib/data-service-clients.js
@@ -2,17 +2,28 @@
 
 import { revalidatePath } from "next/cache";
 import { prisma } from "../_lib/prisma";
+import { auth } from "@/app/_lib/auth";
 
 export async function saveClient(client) {
-  if (!client.id) await prisma.client.create({ data: { name: client.name } });
+  const session = await auth();
+  if (!session) return;
+
+  if (!client.id)
+    await prisma.client.create({
+      data: { name: client.name, user: { connect: { id: session.user.id } } },
+    });
   else
     await prisma.client.update({
       where: { id: client.id },
       data: { name: client.name },
+      user: { connect: { id: session.user.id } },
     });
 }
 
 export async function getClients(skip = 0, take = 10) {
+  const session = await auth();
+  if (!session) return;
+
   try {
     const [clients, totalCount] = await Promise.all([
       prisma.client.findMany({
@@ -21,6 +32,7 @@ export async function getClients(skip = 0, take = 10) {
         orderBy: {
           name: "asc", // Сортировка по имени
         },
+        where: { userId: session.user.id },
       }),
       prisma.client.count(),
     ]);
@@ -40,8 +52,11 @@ export async function getClients(skip = 0, take = 10) {
 }
 
 export async function deleteClient(id) {
+  const session = await auth();
+  if (!session) return;
+
   try {
-    await prisma.client.delete({ where: { id } });
+    await prisma.client.delete({ where: { id, userId: session.user.id } });
     revalidatePath("/clients");
   } catch (error) {
     throw error;
@@ -49,8 +64,11 @@ export async function deleteClient(id) {
 }
 
 export async function getClient(id) {
+  const session = await auth();
+  if (!session) return;
+
   return prisma.client.findUnique({
-    where: { id },
+    where: { id, userId: session.user.id },
   });
 }
