+++ app/_lib/data-service-orders.js
@@ -4,8 +4,12 @@ import { format } from "date-fns";
 import { revalidatePath } from "next/cache";
 import { prisma } from "../_lib/prisma";
 import { dateFormat } from "@/app/_helpers/appConstants";
+import { auth } from "@/app/_lib/auth";
 
 export async function getOrders(skip = 0, take = 10) {
+  const session = await auth();
+  if (!session) return;
+
   try {
     const [orders, totalCount] = await Promise.all([
       prisma.order.findMany({
@@ -22,6 +26,7 @@ export async function getOrders(skip = 0, take = 10) {
         orderBy: {
           date: "desc",
         },
+        where: { userId: session.user.id },
       }),
       prisma.order.count(),
     ]);
@@ -41,11 +46,15 @@ export async function getOrders(skip = 0, take = 10) {
 }
 
 export async function getAvailableOrders(clientId) {
+  const session = await auth();
+  if (!session) return;
+
   try {
     return await prisma.order.findMany({
       where: {
         invoiceId: null,
         clientId: Number(clientId),
+        userId: session.user.id,
       },
       include: {
         client: {
@@ -62,9 +71,12 @@ export async function getAvailableOrders(clientId) {
 }
 
 export async function getOrder(id) {
+  const session = await auth();
+  if (!session) return;
+
   try {
     return prisma.order.findUnique({
-      where: { id },
+      where: { id, userId: session.user.id },
       include: {
         client: {
           select: {
@@ -80,10 +92,13 @@ export async function getOrder(id) {
 }
 
 export async function saveOrder(order) {
+  const session = await auth();
+  if (!session) return;
+
   if (!order.id) {
     try {
       await prisma.order.create({
-        data: order,
+        data: { ...order, user: { connect: { id: session.user.id } } },
       });
     } catch (err) {
       throw err;
@@ -91,7 +106,7 @@ export async function saveOrder(order) {
   } else {
     try {
       await prisma.order.update({
-        where: { id: order.id },
+        where: { id: order.id, userId: session.user.id },
         data: {
           date: order.date,
           amount: order.amount,
@@ -107,8 +122,11 @@ export async function saveOrder(order) {
 }
 
 export async function deleteOrder(id) {
+  const session = await auth();
+  if (!session) return;
+
   try {
-    await prisma.order.delete({ where: { id } });
+    await prisma.order.delete({ where: { id, userId: session.user.id } });
     revalidatePath("/orders");
   } catch (error) {
     throw error;
@@ -129,6 +147,9 @@ export async function getNewOrder() {
 }
 
 export async function getSales() {
+  const session = await auth();
+  if (!session) return;
+
   const currentDate = new Date();
   const thirtyDaysAgo = new Date();
   thirtyDaysAgo.setDate(currentDate.getDate() - 30);
@@ -142,6 +163,7 @@ export async function getSales() {
             gte: thirtyDaysAgo,
             lte: currentDate,
           },
+          userId: session.user.id,
         },
         select: {
           date: true,
@@ -157,6 +179,7 @@ export async function getSales() {
             gte: thirtyDaysAgo,
             lte: currentDate,
           },
+          userId: session.user.id,
         },
         select: {
           date: true,
